% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_models.R
\name{estimate_models}
\alias{estimate_models}
\title{Estimate Bayesian Graphical Models}
\usage{
estimate_models(
  level = c("Gaussian", "Discrete"),
  variable_type = "ordinal",
  reference_category = 1,
  repetitions,
  data,
  no_observations,
  no_variables,
  no_categories,
  interaction_scale = 2.5,
  iter = 10000,
  induce_sparsity = TRUE,
  density = 0.5,
  no_cores = NULL,
  save = FALSE,
  edge_prior = "Bernoulli",
  threshold_alpha = 0.5,
  threshold_beta = 0.5,
  edge_selection = TRUE,
  inclusion_probability = 0.5,
  beta_bernoulli_alpha = 1,
  beta_bernoulli_beta = 1,
  df.prior = 3,
  algorithm = "bdmcmc",
  burnin = 1000
)
}
\arguments{
\item{level}{The type of data. It can be either "Gaussian" or "Discrete". The choice of which model (and package) for model estimation depends on this argument. If \code{level = Gaussian} then the package \code{BDgraph} is used, and if \code{level = Discrete} then the package \code{bgms} is used.}

\item{variable_type}{The type of ordinal variables, if \code{level = "Discrete"}? This can be a single character string specifying the variable type of all p variables at once. Currently, the \code{R} package \code{bgms} supports \code{“ordinal”} and \code{“blume-capel”} variable types. Binary variables are automatically treated as \code{“ordinal”}. Defaults to \code{variable_type = "ordinal"}. For more details see the documentation of the \code{R} package \code{bgms}.}

\item{reference_category}{An integer vector of length \code{no_variables} specifying the \code{reference_category} that is used in case \code{variable_type = 'blume-capel'}. Can be any integer value between 0 and no_categories (or \code{no_categories[i]}).}

\item{repetitions}{The number of repetitions in the simulation.}

\item{data}{The simulated data sets.}

\item{no_observations}{The sample size of the simulated datasets.}

\item{no_variables}{The number of variables in the simulated data.}

\item{no_categories}{The number of ordinal categories (only relevant if level = "Discrete").}

\item{interaction_scale}{The interaction_scale parameter for the interaction prior (only relevant if level = "Discrete").}

\item{iter}{The number of iterations for the estimation algorithm.}

\item{induce_sparsity}{If TRUE then it induces sparsity in the network structure based on the values provided in the \code{density} argument.}

\item{density}{The density/sparsity of the adjacency matrix used to generate the data sets (only relevant if \code{induce_sparsity = TRUE}).}

\item{no_cores}{The number of CPU cores to use for parallel computation. The parallel computation is only used for the Bayesian analysis, and not for simulating the data. Defaults to the maximum number of cores available on the user's machine.}

\item{save}{If TRUE, the samples fro the MCMC algorithm are saved from all the iterations \code{iter}. Only possible when level = "Discrete".}

\item{edge_prior}{The prior distribution for the edge parameters (only relevant if level = "Discrete").}

\item{threshold_alpha, threshold_beta}{The shape parameters of the beta-prime prior density for the threshold parameters. Must be positive values. If the two values are equal, the prior density is symmetric about zero. If threshold_beta is greater than threshold_alpha, the distribution is skewed to the left, and if threshold_beta is less than threshold_alpha, it is skewed to the right. Smaller values tend to lead to more diffuse prior distributions.}

\item{edge_selection}{Should the function perform Bayesian edge selection on the edges of the MRF in addition to estimating its parameters (edge_selection = TRUE), or should it just estimate the parameters (edge_selection = FALSE)? The default is edge_selection = TRUE.}

\item{inclusion_probability}{The prior edge inclusion probability for the Bernoulli model. Can be a single probability, or a matrix of p rows and p columns specifying an inclusion probability for each edge pair. The default is inclusion_probability = 0.5. Corresponds to the g.prior argument in the BDgraph package.}

\item{beta_bernoulli_alpha, beta_bernoulli_beta}{The two shape parameters of the Beta prior density for the Bernoulli inclusion probability. Must be positive numbers. Defaults to beta_bernoulli_alpha = 1 and beta_bernoulli_beta = 1.}

\item{df.prior}{The degrees of freedom for the prior distribution of the precision matrix (only relevant if level = "Gaussian").}

\item{algorithm}{The algorithm to use for estimating the Gaussian Graphical Model (only relevant if level = "Gaussian").}

\item{burnin}{The number of burn-in iterations for the MCMC algorithm}
}
\value{
A list containing the estimated Bayesian Graphical Models and the associated parameter gird.
}
\description{
This function estimates Bayesian Graphical Models (BGM) using either the \code{bgms} or \code{BDGraph} R packages, depending on the level of measurement specified. It supports parallelization for efficient computation on multiple CPU cores.
}
\details{
This function estimates BGMs using either the \code{bgms} or \code{BDGraph} R packages, depending on the specified level. It parallelizes the estimation process for efficient computation.
}
