% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{Generate Data from a Markov Random Field Model}
\usage{
generate_data(
  level = c("Gaussian", "Discrete"),
  variable_type = "ordinal",
  reference_category,
  repetitions,
  no_observations,
  no_variables,
  no_categories,
  induce_sparsity = TRUE,
  density = 0.5,
  irt = TRUE,
  dataset,
  bc_alpha,
  bc_beta
)
}
\arguments{
\item{level}{Character vector specifying the type of data to generate. It can be either "Gaussian" or "Discrete".}

\item{variable_type}{What kind of variables are there? Can be a single character string specifying the variable type of all p variables at once or a vector of character strings of length p specifying the type for each variable in x separately. Currently, bgm supports “ordinal” and “blume-capel”. Binary variables are automatically treated as “ordinal’’. Defaults to variable_type = "ordinal". Only relevant if level = "Discrete".}

\item{reference_category}{An integer vector of length no_variables specifying the reference_category category that is used for the Blume-Capel model (details below). Can be any integer value between 0 and no_categories (or \code{no_categories[i]}).}

\item{repetitions}{Numeric. The number of repetitions in the simulation.}

\item{no_observations}{Numeric vector. The sample size of the simulated datasets.}

\item{no_variables}{Numeric vector. The number of variables in the simulated data.}

\item{no_categories}{Numeric vector. The number of ordinal categories (only relevant if level = "Discrete").}

\item{induce_sparsity}{Logical. If TRUE then it induces sparsity in the network structure based on the values provided in the \code{density} argument.}

\item{density}{Numeric vector. The density/sparsity of the adjacency matrix used to generate the data sets (only relevant if \code{induce_sparsity = TRUE}).}

\item{irt}{Logical. If TRUE, an IRT model is used to generate the maximum pseudolikelihood parameters used to generate the simulated data sets (defaults to TRUE).}

\item{dataset}{Data frame. In case level = "Discrete" and irt = FALSE, the user can provide a dataset which will be used to generate the maximum pseudolikelihood parameters. Note that the number of variables in the provided data set needs to be equal to or larger than the highest number provided in the argument \code{no_variables}. Also, the number of ordinal categories in the provided data set must be larger than or equal to the number of ordinal categories provided in \code{no_categories}.}

\item{bc_alpha, bc_beta}{Numeric. bc_alpha the linear contribution of the Blume-Capel model andbc_beta the quadratic contribution. Defaults to 0.5 and 0.5, respectively.}
}
\value{
A list containing the generated data, maximum pseudolikelihood estimates (MPLEs) used to generate the data sets, and the adjacency matrices based on the provided required density through the density argument.
}
\description{
This function generates data from a Markov Random Field (MRF) model, which can either be Gaussian or Discrete. For the Discrete case, the function can generate data from an Item Response Theory (IRT) model if specified. The generated data can be used for various simulation studies.
}
\details{
The parameters no_observations, no_variables, and no_categories can be specified as vectors to generate data for multiple scenarios in one function call.
}
\examples{
# Generate Ordinal data sets using parameters from an IRT model
generate_data(level = "Discrete",
              variable_type = "ordinal",
              repetitions  = 3,
              no_observations = 200,
              no_variables = 15,
              no_categories = 4,
              induce_sparsity = TRUE,
              density = 0.1,
              irt = TRUE)
}
