% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simBgms-package.R
\docType{package}
\name{simBgms-package}
\alias{simBgms}
\alias{simBgms-package}
\title{simBgms: Simplifying Simulation Studies for Bayesian Markov Random Field Models}
\description{
The \code{R} package \strong{simBgms} provides an easy way to perform
simulation studies for (Bayesian) graphical modeling. This package can be used to simulate data for Markov random field models
and estimate those models using the \code{R} packages \strong{bgms} and \strong{BDgraph}. It provides
an easy way to perform simulation studies without requiring advanced programming skills.
The package can parallel the estimation of the models, making the process
computationally efficient.
}
\author{
\strong{Maintainer}: Nikola Sekulovski \email{n.sekulovski@uva.nl} (\href{https://orcid.org/orcid.org/0000-0001-7032-1684}{ORCID})

}
\keyword{internal}
